+++ qom/numerics/__init__.py
@@ -1,0 +1,1 @@
+from qom.numerics.RHCriterion import RHCriterion